+++ autoblocks/_impl/prompts/autogenerate.py
@@ -1,8 +1,6 @@
-import re
 from typing import Any
 from typing import Dict
 from typing import List
-from typing import Optional
 
 import httpx
 
@@ -13,6 +11,10 @@ from autoblocks._impl.prompts.models import AutogeneratePromptsConfig
 from autoblocks._impl.prompts.models import FrozenModel
 from autoblocks._impl.prompts.placeholders import TemplatePlaceholder
 from autoblocks._impl.prompts.placeholders import parse_placeholders_from_template
+from autoblocks._impl.prompts.utils import indent
+from autoblocks._impl.prompts.utils import infer_type
+from autoblocks._impl.prompts.utils import to_snake_case
+from autoblocks._impl.prompts.utils import to_title_case
 from autoblocks._impl.util import AutoblocksEnvVar
 from autoblocks._impl.util import encode_uri_component
 
@@ -84,57 +86,6 @@ from autoblocks.prompts.renderer import TemplateRenderer
 from autoblocks.prompts.renderer import ToolRenderer
 
 """
-
-
-def to_title_case(s: str) -> str:
-    # Remove leading numbers
-    s = re.sub(r"^\d+", "", s)
-    # Replace all non-alphanumeric characters with underscores
-    s = re.sub(r"[^a-zA-Z0-9]+", "_", s)
-    # Replace all underscores with capital letter of next word
-    return re.sub(r"(?:^|_)(.)", lambda m: m.group(1).upper(), s)
-
-
-def to_snake_case(s: str) -> str:
-    # Remove leading numbers
-    s = re.sub(r"^\d+", "", s)
-
-    # Replace all non-alphanumeric characters with spaces
-    s = re.sub(r"[^a-zA-Z0-9]+", " ", s)
-
-    # Replace spaces with underscores
-    s = re.sub(r"\s+", "_", s)
-
-    # Add underscores between camel case words
-    s = re.sub(r"([A-Z]+)([A-Z][a-z])", r"\1_\2", s)
-    s = re.sub(r"([a-z\d])([A-Z])", r"\1_\2", s)
-
-    # Remove leading and trailing underscores
-    s = s.strip("_")
-
-    return s.lower()
-
-
-def infer_type(value: Any) -> Optional[str]:
-    if isinstance(value, str):
-        return "str"
-    elif isinstance(value, bool):
-        return "bool"
-    elif isinstance(value, (int, float)):
-        # The default union mode in pydantic is "smart" mode,
-        # which will use the most specific type possible.
-        # https://docs.pydantic.dev/latest/concepts/unions/
-        return "Union[float, int]"
-    elif isinstance(value, list):
-        if len(value) > 0:
-            return f"list[{infer_type(value[0])}]"
-    elif isinstance(value, dict):
-        return "Dict[str, Any]"
-    return None
-
-
-def indent(times: int = 1, size: int = 4) -> str:
-    return " " * size * times
 
 
 def generate_params_class_code(prompt: PromptCodegen) -> str:
